adding base user feature
