+++ src/routes/user.routes.js
@@ -1,0 +1,7 @@
+const express = require('express')
+const router = express.Router()
+const userController = require('../controller/user.controller')
+
+router.get('/', userController.getAll);
+
+module.exports = router;
