+++ src/service/user.service.js
@@ -1,0 +1,10 @@
+// regras de negocio e validações
+const userRepository = require('../repository/user.repository');
+
+
+function getAll(){
+    const users = userRepository.getAll();
+    return users;
+};
+
+module.exports = { getAll};
